Use Header Files From swift-corelibs-foundation Project
While I was looking at Apple's SDK, I noticed that some of the headers files in CoreFoundation are from the
swift-corelibs-foundation project. So I decided to symbol link the header files that exist in that repo.

Interestingly, there are two header files (CFCGTypes.h and CFXMLNode.h) that are licensed under Apache License v2.0,
but are not included in the swift-corelibs-foundation. Not sure why Apple didn't include those files in the repo...